Add an AWS Budget for this service
Keep costs under close supervision with an AWS Budget and notification.
The budgeted amount was based on the [ngrok.io pricing for the _Pro_
tier](https://ngrok.com/pricing) since that's most comparable in terms
of final features.

Filters costs based on a customr _Service_ resource tag and sends an
alert at 80% of forecasted monthly budget.